LPS-77714 Make ServiceTrackerCollectionImpl a static inner class to remove internal package in registry-api.
